// src/app/models/contrat.ts
export class contrat {

    idContrat : number=0;
    nomClient: string='';
    imageContrat: string='';
    dateCreation: string='';
    // dateCreation: string=''; 
    // typeContrat: string='';
    // nomAcheteur: string='';
    // nomVendeur: string='';
    // objetContrat:number=0;
    // prixAchatVente: string='';
    // confidentialite: string='';
    // ModalitePaiement: string='';
    // imageContrat: string='';

    
}